services/auth_service/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
drf-spectacular>=0.27
djangorestframework-simplejwt>=5.3
dj-database-url>=2.1
psycopg2-binary>=2.9
redis>=5.0
celery>=5.3
pytest>=7.4
pytest-django>=4.8

